fixing some norms
