+++ src/execute/redir_utils.c
