+++ src/handle/env_var_utils.c
@@ -6,7 +6,7 @@
 /*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/26 18:44:05 by mely-pan          #+#    #+#             */
-/*   Updated: 2025/04/13 17:12:48 by htrindad         ###   ########.fr       */
+/*   Updated: 2025/05/31 20:03:13 by mely-pan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,16 +25,69 @@ char	*var_cases(t_ms *shell, int *i)
 	return (tmp);
 }
 
-char	*conc_char(char c)
+bool	has_env_var(const char *s)
 {
-	char *tmp;
-	char str[2];
+	int	i;
 
-	tmp = ft_strdup("");
+	i = 0;
+	while (s[i])
+	{
+		if (s[i] == '\'')
+		{
+			i++;
+			while (s[i] && s[i] != '\'')
+				i++;
+		}
+		else if (s[i] == '$')
+		{
+			i++;
+			if (s[i] == '?')
+				return (true);
+			if (ft_isalpha(s[i]) || s[i] == '_')
+				return (true);
+		}
+		i++;
+	}
+	return (false);
+}
+
+char	*get_env_value(t_env *env, char *env_var)
+{
+	t_env	*tmp;
+	char	*env_value;
+
+	tmp = env;
+	while (tmp)
+	{
+		if (tmp->key && !ft_strncmp(tmp->key, env_var, ft_strlen(env_var)))
+		{
+			env_value = ft_strdup(tmp->value);
+			if (!env_value)
+				return (NULL);
+			return (env_value);
+		}
+		tmp = tmp->next;
+	}
+	return (NULL);
+}
+
+char	*extract_env_var(t_ms *shell, int *i)
+{
+	char	*env_var;
+	char	*tmp;
+	int		j;
+
+	j = *i;
+	while (shell->input[j] && (ft_isalnum(shell->input[j])
+			|| shell->input[j] == '_'))
+		j++;
+	env_var = ft_substr(shell->input, *i, j - *i);
+	if (!env_var)
+		return (NULL);
+	tmp = get_env_value(shell->env, env_var);
+	free(env_var);
 	if (!tmp)
 		return (NULL);
-	str[0] = c;
-	str[1] = '\0';
-	ft_strjoin(tmp, str);
+	*i = j - 1;
 	return (tmp);
 }
